requirements.txt
sanic==23.12.1
sanic-ext==23.12.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.1.2
uvicorn==0.25.0
python-multipart==0.0.6